services/intelli-search/data-pipeline/requirements.txt
# Core dependencies
pandas==2.1.3
numpy==1.26.2
opensearch-py==2.3.1
python-dotenv==1.0.0
pyyaml==6.0.1
structlog>=24.0.0
sentence-transformers==2.7.0
huggingface-hub==0.25.2
torch>=2.0.0
boto3>=1.34.0
pytest==7.4.3
pytest-cov==4.1.0
